requirements.txt
streamlit
opencv-python
pillow
numpy
matplotlib